refactor(cli): split runtime output builders
