+++ .sampo/changesets/838-split-runtime-output.md
@@ -1,0 +1,5 @@
+---
+npm/wp-typia: patch
+---
+
+Split runtime completion payload builders into command-focused modules while preserving the existing runtime-bridge-output facade.
